+++ src/components/UI/index.ts
@@ -1,3 +1,5 @@
 export * from './Button';
 export * from './Input';
+export * from './Modal';
 export * from './Select';
+export * from './Link';
